requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.30.1
gunicorn==22.0.0
SQLAlchemy==2.0.31
alembic==1.13.2
psycopg2-binary==2.9.9
asyncpg==0.29.0
python-dotenv==1.0.1
pydantic==2.7.4
pydantic-settings==2.3.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
email-validator==2.2.0
httpx==0.27.0
requests==2.32.3
python-multipart==0.0.9
greenlet==3.0.3
pytest==8.3.3
pytest-asyncio==0.23.7
pytest-cov==5.0.0
aiosqlite==0.20.0

